Python/APP1_python_bootcamp_1/Python_Bootcamp.Day_08-2/src/requirements.txt
httpx
"fastapi[standard]"
pydantic
uvicorn
aioredis